backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.12
psycopg2-binary>=2.9.10
anthropic>=0.85.0
plaid-python>=38.4.0
pandas>=2.3.0
numpy>=2.4.0
rapidfuzz>=3.14.0
python-dotenv>=1.2.0
cryptography>=43.0.0
bcrypt>=4.2.0
itsdangerous>=2.2.0
slowapi>=0.1.9
